Module Peahell

val print_parens : ?⁠max_level:int -> ?⁠at_level:int -> Stdlib.Format.formatter -> ('a, Stdlib.Format.formatter, unit, unit) Stdlib.format4 -> 'a

Print an expression, possibly placing parentheses around it. We always print things at a given "level" at_level. If the level exceeds the maximum allowed level max_level then the expression should be parenthesized.

Let us consider an example. When printing nested applications, we should print App (App (e1, e2), e3) as "e1 e2 e3" and App(e1, App(e2, e3)) as "e1 (e2 e3)". So if we assign level 1 to applications, then during printing of App (e1, e2) we should print e1 at max_level 1 and e2 at max_level 0.

module Location : sig ... end

module Report : sig ... end

module Input : sig ... end

type filename = Peahell__.Language.filename

module type Language = Peahell__.Language.S

The definition of a programming language

module Make : functor (L : Language) -> sig ... end

Create a language from its definition.