Module Peahell_native.Make

Parameters

module L : Peahell.Language

Signature

module History : sig ... end

type mode =

| Batch

| Interactive

| Expect

Should the interactive shell be run?

val mode : mode Stdlib.ref

type trace_out =

| Stdout

| File of string

Trace printing

val trace_out : trace_out Stdlib.ref

val usage : string

The usage message.

val files : string list Stdlib.ref

A list of files to be loaded and run.

val add_file : string -> unit

Add a file to the list of files to be loaded, and record whether it should be processed in interactive mode.

val options : (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list

Command-line options

val anonymous : string -> unit

Treat anonymous arguments as files to be run.

val read_file : (string -> Stdlib.Lexing.lexbuf -> 'a) -> string -> 'a

Parse the contents from a file, using a given parser.

val read_toplevel : (Stdlib.Lexing.lexbuf -> 'a) -> unit -> 'a

Parse input from toplevel, using the given parser.

val use_file : L.environment -> string -> L.environment

Load directives from the given file.

val batch : L.environment -> string list -> L.environment

val expect : L.environment -> string -> unit

Expect mode

val toplevel : L.environment -> unit

Interactive toplevel

val run : (unit -> unit) -> 'a

val main : unit -> 'a

Main program